Fix load data bug
All tasks will be stored as ToDo and DateTime object is stored in a form not handled by the read function.

These are fundamental errors that must be resolved.

Let's ensure the getStoreFormat functions for Deadline and Event are correctly identified with the right type marker (T/D/E) and format the dateTime to that of the user input.

Storing dateTime in the format of a user input will allow for code reuse of Parser.parseDate() when we read the data file.